#8151: Move sweep test
