+++ ttnn/cpp/ttnn/operations/eltwise/unary/unary_pybind.hpp
@@ -433,7 +433,8 @@ void bind_unary_operation_with_dim_parameter(
     const unary_operation_t& operation,
     const std::string& parameter_name,
     const std::string& parameter_doc,
-    const std::string& info_doc) {
+    const std::string& info_doc,
+    const std::string& note = "") {
 
     auto doc = fmt::format(
         R"doc(
@@ -452,6 +453,9 @@ void bind_unary_operation_with_dim_parameter(
         Returns:
             ttnn.Tensor: the output tensor.
 
+        Note:
+            {5}
+
         Example:
             >>> tensor = ttnn.from_torch(torch.tensor((1, 2), dtype=torch.bfloat16), device=device)
             >>> output = {1}(tensor, {2})
@@ -460,7 +464,8 @@ void bind_unary_operation_with_dim_parameter(
         operation.python_fully_qualified_name(),
         parameter_name,
         parameter_doc,
-        info_doc);
+        info_doc,
+        note);
 
     bind_registered_operation(
         module,
@@ -1628,13 +1633,15 @@ void py_module(py::module& module) {
     // Unary ops with dim parameter
     detail::bind_unary_operation_with_dim_parameter(module, ttnn::glu, "dim", "Dimenstion to split input tensor. Supported dimension -1 or 3", "Split the tensor into two, apply glu function on second tensor followed by mul op with first tensor");
     detail::bind_unary_operation_with_dim_parameter(module, ttnn::reglu, "dim", "Dimenstion to split input tensor. Supported dimension -1 or 3", "Split the tensor into two, apply relu function on second tensor followed by mul op with first tensor",
-    R"doc(Supported dtypes, layouts, and ranks:
+        R"doc(Supported dtypes, layouts, and ranks:
 
-        +----------------------------+---------------------------------+-------------------+
-        |     Dtypes                 |         Layouts                 |     Ranks         |
-        +----------------------------+---------------------------------+-------------------+
-        |    BFLOAT16, BFLOAT8_B     |          TILE                   |      2, 3, 4      |
-        +----------------------------+---------------------------------+-------------------+)doc");
+           +----------------------------+---------------------------------+-------------------+
+           |     Dtypes                 |         Layouts                 |     Ranks         |
+           +----------------------------+---------------------------------+-------------------+
+           |    BFLOAT16, BFLOAT8_B     |          TILE                   |      2, 3, 4      |
+           +----------------------------+---------------------------------+-------------------+
+        )doc");
+
     detail::bind_unary_operation_with_dim_parameter(module, ttnn::geglu, "dim", "Dimenstion to split input tensor. Supported dimension -1 or 3", "Split the tensor into two, apply gelu function on second tensor followed by mul op with first tensor");
     detail::bind_unary_operation_with_dim_parameter(module, ttnn::swiglu, "dim", "Dimenstion to split input tensor. Supported dimension -1 or 3", "Split the tensor into two, apply silu function on second tensor followed by mul op with first tensor");
